Remove file generation and linter from `unit-test` makefile target.
Golang test has it own internal linter which will fail the UTs if there
are some basic lint errors and having that linter overhead when writing
UTs and continuously running them is a pain.

Same apply for files generation. If mocks and the CRD aren't up to date,
UTs will fail anyway.

All those targets are tested in CI anyway so we won't forget it, this is
just a matter of convenient when writing UTs.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>