Add permissions to module loader pod
To be able to load a module, more privileges are required.
This change sets the SELinux type of the pod to `spc_t` and the user to
`root`. It also explicitely disable privilege escalation.

This has been tested with a pod ServiceAccount that can use the
`privileged` SCC.

Reference: [Introducing a Super Privileged Container
Concept](https://developers.redhat.com/blog/2014/11/06/introducing-a-super-privileged-container-concept)

Signed-off-by: Fabien Dupont <fdupont@redhat.com>