Load firmware from the module loader
It is common for kernel module to come with firmware(s). On immutable
operating systems like CoreOS, the standard firmware path
`/lib/firmware` is read-only. So, we need to let the kernel know where
to lookup firmware file(s).

This change adds the `FirmwarePath` attribute to the `ModprobeSpec`
definition and modify the `MakeLoadCommand` and `MakeUnloadCommand`
methods to leverage it.

It also adds a `Volume` and a `VolumeMount` to expose the
`/var/lib/firmware` host directory (created if it doesn't exist) inside
the pod. The `/var/lib/firmware` is [the recommended path for firmwares
in OpenShift with RHCOS
nodes](https://access.redhat.com/documentation/fr-fr/openshift_container_platform/4.10/html-single/post-installation_configuration/index#rhcos-load-firmware-blobs_post-install-machine-configuration-tasks).

The `MakeLoadCommand` adds a step to copy the `FirmwarePath` to
`/var/lib/firmware` to the `modprobe` command. The
`MakeUnloadCommand` removes the firmware files previously copied
after unloading the module. The commands use `/usr/bin -c` to allow
shell control like `&&`.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>